requirements.txt
langgraph
langchain-core
langchain-openai
openai
python-dotenv
sqlalchemy
pymysql
streamlit
pandas